LC-sqrtx [Runtime: 0 ms; Memory: 8.6 MB]

+++ binary-search/sqrtx.cpp
@@ -2,9 +2,9 @@ class Solution {
 public:
     int mySqrt(int x) {
         if(x < 1) return 0;
-        int left = 1, right = x;
+        long long left = 1, right = x;
         while(left < right){
-            long long mid = (long long) (left + right  + 1) / 2;
+            long long mid = (left + right  + 1) / 2;
             if(mid * mid > x) right = mid -1 ;
             else left = mid;
         }
